+++ NEWS.md
@@ -111,6 +111,7 @@
 - Method `$over()` gains an argument `mapping_strategy` (#984, #988).
 - New method `$item()` for `DataFrame` and `Series` (#992).
 - New active binding `<Series>$struct$fields` (#1002).
+- New method `$clear()` for `DataFrame`, `LazyFrame`, and `Series` (#1004).
 
 ### Bug fixes